requirements.txt
requests
python-dotenv
schedule
prometheus-client
python-dxf
jira